requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pydantic>=2.0
python-dotenv>=1.0
typing-extensions